liquibase update,ручка fullUpdateBicDirectoryEntry + fixiki

+++ central-bank-rest-api/src/main/java/ru/bluewater/centralbankrestapi/api/dto/request/create/PartInfoCreateRequestDTO.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.UUID;
 
 @Getter
@@ -12,5 +13,5 @@ public class PartInfoCreateRequestDTO {
     private UUID ed807Uuid;
     private Integer partNo;
     private Integer partQuantity;
-    private BigDecimal partAggregateID;
+    private BigInteger partAggregateID;
 }
