+++ central-bank-rest-api/src/main/java/ru/bluewater/centralbankrestapi/api/dto/request/update/PartInfoUpdateRequestDTO.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.UUID;
 
 @Getter
@@ -11,5 +12,5 @@ import java.util.UUID;
 public class PartInfoUpdateRequestDTO {
     private Integer partNo;
     private Integer partQuantity;
-    private BigDecimal partAggregateID;
+    private BigInteger partAggregateID;
 }
